debugging translatedb

+++ src-hlr-service/admin/hlrTranslateDb.cpp
@@ -1,4 +1,4 @@
-//$Id: hlrTranslateDb.cpp,v 1.1.2.1.4.66 2011/11/28 10:31:20 aguarise Exp $
+//$Id: hlrTranslateDb.cpp,v 1.1.2.1.4.67 2011/11/28 12:56:59 aguarise Exp $
 // -------------------------------------------------------------------------
 // Copyright (c) 2001-2002, The DataGrid project, INFN, 
 // All rights reserved. See LICENSE file for details.
@@ -373,6 +373,29 @@ int upgrade_R_4_0_0(database& DB)
 					time0 = time(NULL);
 				}
 			}
+
+		}
+		long int fromRecordNum = 0;
+		long int toRecordNum = -1;
+		upgradeQuery = "SELECT count(*) from jobTransSummary";
+		if ( debug )
+		{
+			cerr << upgradeQuery << endl;
+		}
+		hlrGenericQuery check1(upgradeQuery);
+		check1.query();
+		if ( check1.errNo == 0)
+		{
+			long int fromRecordNum = atoi((((check1.queryResult).front())[1]).c_str());
+		}
+		hlrGenericQuery check2(upgradeQuery);
+		check2.query();
+		if ( check2.errNo == 0)
+		{
+			long int toRecordNum = atoi((((check1.queryResult).front())[1]).c_str());
+		}
+		if ( fromRecordNum == toRecordNum )
+		{
 			upgradeQuery = "DROP TABLE jobTransSummary";
 			hlrGenericQuery upgrade5(upgradeQuery);
 			upgrade5.query();
@@ -407,6 +430,12 @@ int upgrade_R_4_0_0(database& DB)
 				}
 			}
 		}
+		else
+		{
+			cerr << "There was a problem in the UPGRADE phase. Please contact dgas-suuprt@to.infn.it" << endl;
+			res = 2;
+		}
+
 	}
 	return res;
 }
